Refactored report generation logic to aggregate latency distribution histograms with constituent reports. Added pylint and gtest to ci test script
