+++ scripts/lib/xpedite/jupyter/result.py
@@ -9,22 +9,6 @@ Author: Manikandan Dhamodharan, Morgan Stanley
 
 import zlib
 import base64
-import logging
-
-LOGGER = logging.getLogger(__name__)
-
-class Flot(object):
-  """Class to store data for creating latency distribution visualizations"""
-
-  def __init__(self, name, description, data, statistics, width, height, options, checked):
-    self.name = name
-    self.description = description
-    self.data = data
-    self.statistics = statistics
-    self.width = width
-    self.height = height
-    self.options = options
-    self.checked = checked
 
 class Report(object):
   """Class to store compressed html reports"""
@@ -48,51 +32,33 @@ class Result(object):
     self.reportCells = []
     self.envReport = None
 
-  def __getattr__(self, attr):
-    return None
+  @staticmethod
+  def _buildReport(title, description, content):
+    zContent = zlib.compress(content)
+    zContent = base64.b64encode(zContent)
+    return Report(title, description, zContent)
 
-  def attachXpediteReport(self, _, title, description, content):
+  def attachEnvReport(self, title, description, content):
     """
-    Attaches a html report to this result object
+    Attaches a html report with details about the test environment
 
     :param description: description of the report
     :param title: title of the report
     :param content: html content of the report
 
     """
-    zContent = zlib.compress(content)
-    zContent = base64.b64encode(zContent)
-    report = Report(title, description, zContent)
-    if title == 'Test Environment Report':
-      self.envReport = report
-    else:
-      try:
-        self.reportCells[len(self.reportCells) - 1].htmlReport.append(report)
-      except IndexError as ex:
-        LOGGER.exception('No report data generated by the Profiler')
-        raise ex
+    self.envReport = self._buildReport(title, description, content)
 
-  def flot(self, name, description, data, options, statistics=None, width=1000, height=500, checked=None):
+  def attach(self, report):
     """
     Adds a new latency distribution visaulization to this result object
-
-    :param name: Category of the histogram
-    :param description: Description of the histogram
-    :param width: Width of the histogram (Default value = 1000)
-    :param height: Height of the histogram (Default value = 500)
-
     """
-    flot = Flot(name, description, data, statistics, width, height, options, checked)
-    reportcell = Reportcell(flot)
+    reportcell = Reportcell(report.histogram)
+    for constituent in report.constituents:
+      report = self._buildReport(constituent.title, constituent.description, constituent.report)
+      reportcell.htmlReport.append(report)
     self.reportCells.append(reportcell)
 
-  def commitXpediteReport(self, app, profiles, reportName):
+  def commit(self, app, profiles, reportName):
     """Commits results"""
     pass
-
-  @staticmethod
-  def le(reference, _=None):
-    """less than or equals comparator"""
-    def _func(value, name=None, *args, **kwargs): # pylint: disable=unused-argument
-      return value <= reference
-    return _func
